Move hardware compatibility checking from SGInterface and make it an independent function
